Shuffle downloads; make them resumable
Downloads are shuffled in order to make better use of your bandwidth if one server is slow, as well as to reduce the chance of your getting banned for hitting one server too hard.  It also improves coverage if several people start downloading but none of them has enough machine resources to download everything.

Downloads are now resumable because we all like to turn off our computers sometimes.